Move reexports and utils to separate modules
